+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 * Use SF Symbols as icon images, [#41](https://github.com/bcylin/QuickTableViewController/pull/41) by [@ezfe](https://github.com/ezfe)
+* Make `init(style:)` a designated initializer
 
 ## v1.2.4